Prefix with "http://" only for xdg_open#open_url(), map this to gX
Also print the command that's run, because it can be confusing if
nothing happens (did it run at all? did xdg_open do nothing?) and add
xdg_open_silent setting to disable this.

Can improve on that by running the command in a job rather than as a
shell command with `&` attached; that way we can check the error code,
but I don't feel like doing that right now.

Fixes #5